refactor(runner): compose the address-to-id seam from the read path
`idStringForEntityAddress` recognized a bare tagged hash with a regex of its
own, beside the recognition `FabricHash` already owns and `entityIdFrom` already
performs. Two readings of one format can part: a hash spelling accepted there
and not here would normalize to a different identity than a read of the same
address reaches, and no case over a table of today's spellings can see that
coming, since the spelling that parts would be one the table does not list.

There is nothing left to recognize. The answer is `toURI(entityIdFrom(address))`
— the pair `Runtime.getCellFromEntityId` puts in front of every read — and an
address that pair refuses comes back as it was. That refusal is not an error
here: an id under another subject's scheme is already an id, a kinded id names
an entity of its own, and a string that addresses nothing has nothing to
normalize.

Composing it rather than deciding it again also makes the seam exact where a
recognizer could only approximate. A padded `fid1:AA==` now yields the id its
canonical spelling names, because that is the document a read of it reaches.

The function moves to `create-ref.ts`, beside the `entityIdFrom` it calls, which
leaves `entity-kind.ts` the import-free leaf that `shell` and `ui` reach for.

The exception the composition takes on an id it cannot address costs 2.7µs,
measured, against the storage round trip on the next line of every caller —
`entityIdExists` per `cd`, and per FUSE path lookup.
