+++ packages/runner/src/index.ts
@@ -79,6 +79,7 @@ export {
   type EntityId,
   entityIdFrom,
   getEntityId,
+  idStringForEntityAddress,
 } from "./create-ref.ts";
 export type { CellResult as QueryResult } from "./query-result-proxy.ts";
 export type {
